+++ lib/Controller/ShiftsTypeController.php
@@ -1,11 +1,16 @@
 <?php
 /*
  * @copyright Copyright (c) 2021. Fabian Kirchesch <fabian.kirchesch@csoc.de>
+ * @copyright Copyright (c) 2023. Kevin Küchler <kevin.kuechler@csoc.de>
  *
  * @author Fabian Kirchesch <fabian.kirchesch@csoc.de>
+ * @author Kevin Küchler <kevin.kuechler@csoc.de>
  */
 
 namespace OCA\Shifts\Controller;
+
+use OCP\AppFramework\Http;
+use OCA\Shifts\Service\PermissionService;
 
 use OCA\Shifts\AppInfo\Application;
 use OCA\Shifts\Service\ShiftsTypeService;
@@ -17,16 +22,16 @@ class ShiftsTypeController extends Controller{
 	/** @var ShiftsTypeService */
 	private $service;
 
-	/** @var string */
-	private $userId;
+	/** @var PermissionService */
+	private $permService;
 
 	use Errors;
 
 
-	public function __construct(IRequest $request, ShiftsTypeService $service, $userId){
+	public function __construct(IRequest $request, ShiftsTypeService $service, PermissionService $permService){
 		parent::__construct(Application::APP_ID, $request);
 		$this->service = $service;
-		$this->userId = $userId;
+		$this->permService = $permService;
 	}
 
 	/**
@@ -70,9 +75,13 @@ class ShiftsTypeController extends Controller{
 	 * @return DataResponse
 	 */
 	public function create(string $name, string $description, string $startTimestamp, string $stopTimestamp, string $color,
-						   int $moRule, int $tuRule, int $weRule, int $thRule, int $frRule, int $saRule, int $soRule, int $skillGroupId, bool $isWeekly, bool $deleted): DataResponse {
+						   int $moRule, int $tuRule, int $weRule, int $thRule, int $frRule, int $saRule, int $soRule, int $skillGroupId, bool $isWeekly): DataResponse {
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return new DataResponse($this->service->create($name, $description, $startTimestamp, $stopTimestamp, $color,
-														$moRule, $tuRule, $weRule, $thRule, $frRule, $saRule, $soRule, $skillGroupId, $isWeekly, $deleted));
+														$moRule, $tuRule, $weRule, $thRule, $frRule, $saRule, $soRule, $skillGroupId, $isWeekly));
 	}
 
 	/**
@@ -99,6 +108,10 @@ class ShiftsTypeController extends Controller{
 	public function update(int $id, string $name, string $desc, string $startTimestamp, string $stopTimestamp, string $color,
 						   int $moRule, int $tuRule, int $weRule, int $thRule, int $frRule, int $saRule, int $soRule, int $skillGroupId, bool $isWeekly, bool $deleted): DataResponse
 	{
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return $this->handleNotFound(function() use ($id, $name, $desc, $startTimestamp, $stopTimestamp, $color,
 			$moRule, $tuRule, $weRule, $thRule, $frRule, $saRule, $soRule, $skillGroupId, $isWeekly, $deleted){
 			return $this->service->update($id, $name, $desc, $startTimestamp, $stopTimestamp, $color,
@@ -114,6 +127,10 @@ class ShiftsTypeController extends Controller{
 	 */
 	public function destroy(int $id): DataResponse
 	{
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return $this->handleNotFound(function() use($id) {
 			return $this->service->delete($id);
 		});
